src/containers/metatask/requirements.txt
celery
requests
biopython